refactor: collection helpers
