+++ packages/starlight/integrations/shared/pathToLocale.ts
@@ -1,6 +1,6 @@
 import type { AstroConfig } from 'astro';
 import type { StarlightConfig } from '../../types';
-import { getCollectionPath } from '../../loaders';
+import { getCollectionPath } from '../../utils/collection';
 import { slugToLocale } from './slugToLocale';
 
 /** Get current locale from the full file path. */
